fix: Dashboard 模块 orgID 类型 uint→int64 及字段名修正
- handler.go: orgID 从 uint 改为 int64，与 middleware 一致
- repository.go: orgID 类型统一为 int64，移除冗余 int64() 转换
- SIRecord: payment_month → start_month 与社保实际字段对齐
- ContractRecord: expiry_date → end_date 与合同模型字段对齐

+++ internal/dashboard/handler.go
@@ -25,7 +25,7 @@ func (h *Handler) GetDashboard(c *gin.Context) {
 		response.Error(c, http.StatusUnauthorized, 40100, "missing org_id in context")
 		return
 	}
-	orgID, ok := orgIDVal.(uint)
+	orgID, ok := orgIDVal.(int64)
 	if !ok {
 		response.Error(c, http.StatusUnauthorized, 40100, "invalid org_id type")
 		return
